:test_tube: test: Ajustar numero dos testes

+++ src/main.c
@@ -170,7 +170,7 @@ int main(){
                     clock_t tempo = clock();
                     buscaPorTextoPat(pat, N, &Memoria);
                     tempo = clock() - tempo;
-                    printf("Tempo de execucao: %f ms\nMemoria alocada: %d bytes\n\n", ((float)(tempo) * 1000)/(CLOCKS_PER_SEC), Memoria);
+                    printf("Tempo de execucao: %lf ms\nMemoria alocada: %d bytes\n\n", ((double)(tempo) * 1000)/(CLOCKS_PER_SEC), Memoria);
                 }
                 else{
                     if(!arquivoEntrada){
